% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-cross.R
\name{contrasts_conversion_check}
\alias{contrasts_conversion_check}
\title{Convert contrasts to a sample-wise label matrix and check it}
\usage{
contrasts_conversion_check(SAMPLES, CONTRASTS, PASS)
}
\arguments{
\item{SAMPLES}{Samples table (samples in rows).}

\item{CONTRASTS}{Contrasts table (group-wise or sample-wise).}

\item{PASS}{boolean. The status of the checks.}
}
\value{
converted contrast df
}
\description{
Convert contrasts to a sample-wise label matrix and check it
}
