% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-table.R
\name{detect_decimal}
\alias{detect_decimal}
\title{Detect decimal separator of text file from its first 100 rows. Assumes
there is a header and rownames column.}
\usage{
detect_decimal(file)
}
\arguments{
\item{file}{Path to a text file.}
}
\value{
"." or ",".
}
\description{
Detect decimal separator of text file from its first 100 rows. Assumes
there is a header and rownames column.
}
