% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{first_feature}
\alias{first_feature}
\title{First name of multi-feature IDs}
\usage{
first_feature(x)
}
\arguments{
\item{x}{Character vector of feature IDs, possibly holding several names
separated by ";", "," or "|".}
}
\value{
Character vector with the first name of each ID.
}
\description{
First name of multi-feature IDs
}
