% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{is.GEO.id.valid}
\alias{is.GEO.id.valid}
\title{Check a GEO accession ID}
\usage{
is.GEO.id.valid(accession)
}
\arguments{
\item{accession}{GEO accession ID.}
}
\value{
TRUE if the ID is non-empty and contains both letters and digits.
}
\description{
Check a GEO accession ID
}
