% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-cross.R
\name{pgx.crosscheckINPUT}
\alias{pgx.crosscheckINPUT}
\title{Cross check input files for pgx.computePGX}
\usage{
pgx.crosscheckINPUT(
  SAMPLES = NULL,
  COUNTS = NULL,
  CONTRASTS = NULL,
  PASS = TRUE
)
}
\arguments{
\item{COUNTS}{data.frame. The data frame corresponding to the input file as in playbase::COUNTS}

\item{CONTRASTS}{data.frame. The data frame corresponding to the input file as in playbase::CONTRASTS}

\item{SAMPLE}{data.frame. The data frame corresponding to the input file as in playbase::SAMPLES}
}
\value{
a list with FIVE elements: SAMPLES, COUNTS and CONTRASTS that are the cleaned version of the
 input data frames, `checks` which contains the status of the checks, and
 `PASS` which contains the overall status of the check.
}
\description{
Cross check input files for pgx.computePGX
}
