% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-counts.R
\name{pgx.getArrayExpress.data}
\alias{pgx.getArrayExpress.data}
\title{Download data from ArrayExpress}
\usage{
pgx.getArrayExpress.data(accession)
}
\arguments{
\item{accession}{ArrayExpress (or GEO) accession ID.}
}
\value{
List with \code{expr}, \code{samples} and \code{source}, or NULL.
}
\description{
Downloads counts and sample metadata for an accession with the
ArrayExpress package. Log2 data is converted to linear.
}
