% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-metadata.R
\name{pgx.getGEOexperimentInfo}
\alias{pgx.getGEOexperimentInfo}
\title{Download the description of a GEO series}
\usage{
pgx.getGEOexperimentInfo(id)
}
\arguments{
\item{id}{GEO accession ID.}
}
\value{
The GSE header (a named list: title, summary, design, ...), or NULL.
}
\description{
Download the description of a GEO series
}
