% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-metadata.R
\name{pgx.getGEOmetadata.fromEset.helper}
\alias{pgx.getGEOmetadata.fromEset.helper}
\title{Sample metadata from one ExpressionSet}
\usage{
pgx.getGEOmetadata.fromEset.helper(eset)
}
\arguments{
\item{eset}{ExpressionSet from GEOquery.}
}
\value{
Data frame of sample metadata.
}
\description{
Sample metadata from one ExpressionSet
}
