% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-metadata.R
\name{pgx.getGEOmetadata.fromEset}
\alias{pgx.getGEOmetadata.fromEset}
\title{Sample metadata from the GEO Series Matrix}
\usage{
pgx.getGEOmetadata.fromEset(id)
}
\arguments{
\item{id}{GEO accession ID.}
}
\value{
Data frame of sample metadata, or NULL.
}
\description{
Downloads the GSE Series Matrix files and extracts the sample
metadata from their ExpressionSets.
}
