% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-metadata.R
\name{pgx.getGEOmetadata.fromGSM}
\alias{pgx.getGEOmetadata.fromGSM}
\title{Sample metadata from GEO GSM records}
\usage{
pgx.getGEOmetadata.fromGSM(id)
}
\arguments{
\item{id}{GEO accession ID.}
}
\value{
Data frame of sample metadata, or NULL.
}
\description{
Builds the sample table from the individual GSM records of a
series, without the GSE Series Matrix files.
}
