% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase.ingest-package.R
\docType{data}
\name{PGX_CHECKS}
\alias{PGX_CHECKS}
\title{Checks performed by pgx.checkINPUT}
\format{
## `data.frame`
rows are checks, columns are description of the check performed.
}
\usage{
PGX_CHECKS
}
\description{
Checks performed by pgx.checkINPUT
}
\keyword{datasets}
