% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase.ingest-package.R
\docType{package}
\name{playbase.ingest-package}
\alias{playbase.ingest}
\alias{playbase.ingest-package}
\title{playbase.ingest: data ingestion for OmicsPlayground}
\description{
Readers and input validation split out of playbase.
}
\keyword{internal}
