% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-olink.R
\name{read_Olink_NPX}
\alias{read_Olink_NPX}
\title{Read Olink NPX data and create a counts matrix}
\usage{
read_Olink_NPX(NPX_data)
}
\arguments{
\item{NPX_data}{Path to Olink file. Must be standard format as per OlinkAnalyze R package.}
}
\value{
NPX data matrix (features on rows; samples on columns)

Sample metadata matrix (samples on rows; metadata on columns)
}
\description{
Read Olink NPX data and create a counts matrix
}
