% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-pheno.R
\name{title2pheno}
\alias{title2pheno}
\title{Get pheno from title [??]}
\usage{
title2pheno(title, split = NULL, trim = TRUE, summarize = TRUE)
}
\arguments{
\item{title}{character vector with the titles}

\item{split}{delimiter character to split on `c(",", ";", "\\|", "_", " ")`.}

\item{trim}{trim leading and trailing whitespace from each term}

\item{summarize}{summarize the terms by counting the number of occurrences of each term}
}
\description{
Get pheno from title [??]
}
