% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_contrasts}
\alias{validate_contrasts}
\title{Validate contrasts data}
\usage{
validate_contrasts(df)
}
\arguments{
\item{df}{Matrix or data frame, or path to a file read with \code{read.as_matrix()}.}
}
\value{
boolean. true if data is valid
}
\description{
Contrasts data is valid if:
 - no duplicate rows
 - no empty rows
 - no duplicate cols
 - contains only cols with "_vs_" in names
}
