% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_samples}
\alias{validate_samples}
\title{Validate samples data}
\usage{
validate_samples(df)
}
\arguments{
\item{df}{Matrix or data frame, or path to a file read with \code{read.as_matrix()}.}
}
\value{
boolean. true if data is valid
}
\description{
Samples data is valid if:
 - no duplicate rows
 - no empty rows
 - no duplicate cols
 - contains less than the max samples allowed
}
