% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-cross.R
\name{contrasts_conversion_check}
\alias{contrasts_conversion_check}
\title{Convert contrasts for OPG}
\usage{
contrasts_conversion_check(SAMPLES, CONTRASTS, PASS)
}
\arguments{
\item{CONTRASTS}{data.frame. The data frame corresponding to the input file as in playbase::CONTRASTS}

\item{PASS}{boolean. The status of the checks.}

\item{SAMPLE}{data.frame. The data frame corresponding to the input file as in playbase::SAMPLES}
}
\value{
converted contrast df
}
\description{
Convert contrasts for OPG
}
