% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-h5.R
\name{h5.readMatrix}
\alias{h5.readMatrix}
\title{Read (sub-)matrix from HDF5 file}
\usage{
h5.readMatrix(
  h5.file,
  rows = NULL,
  cols = NULL,
  matrixid = "data/matrix",
  rowid = "data/rownames",
  colid = "data/colnames"
)
}
\description{
Read (sub-)matrix from HDF5 file
}
