% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-h5.R
\name{h5.readMatrix}
\alias{h5.readMatrix}
\title{Read (sub-)matrix from HDF5 file}
\usage{
h5.readMatrix(
  h5.file,
  rows = NULL,
  cols = NULL,
  matrixid = "data/matrix",
  rowid = "data/rownames",
  colid = "data/colnames"
)
}
\arguments{
\item{h5.file}{Path to the HDF5 file.}

\item{rows, cols}{Row/column names or integer indices to read; NULL for all.}

\item{matrixid, rowid, colid}{HDF5 paths of the matrix and its row and
column names.}
}
\value{
The requested matrix with dimnames.
}
\description{
Read (sub-)matrix from HDF5 file
}
