% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{pgx.checkINPUT}
\alias{pgx.checkINPUT}
\title{Check an uploaded input table before a pgx object is built}
\usage{
pgx.checkINPUT(df, type = c("SAMPLES", "COUNTS", "EXPRESSION", "CONTRASTS"))
}
\arguments{
\item{df}{data.frame. The data frame corresponding to the input file as
described in Omics Playground documentation}

\item{type}{character. One of "SAMPLES", "COUNTS", "EXPRESSION", "CONTRASTS" - The type
of data file to check.}
}
\value{
a list with two elements: `checks` which contains the status of the checks, and
 `PASS` which contains the overall status of the check.
}
\description{
Check an uploaded input table before a pgx object is built
}
