% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-read.R
\name{read_counts}
\alias{read_counts}
\title{Read counts data from file}
\usage{
read_counts(file, first = FALSE, unique = FALSE, paste_char = "_")
}
\arguments{
\item{file}{string. path to file}

\item{first}{boolean. drop multiple feature names (separated by ;)}

\item{unique}{boolean. make duplicated rows unique by pasting a number}

\item{drop_na_rows}{boolean. drop rows without rownames}
}
\description{
Read counts data from file
}
\details{
This function reads a count matrix using \code{read.as_matrix()},
validates it with \code{validate_counts()}, and optionally converts row names from IDs to
gene symbols.

It removes rows with NA, blank or invalid symbols, and collapses any duplicate symbols by
summing counts across rows.
}
\examples{
\dontrun{
counts <- read_counts(playbase.ingest::example_file("counts.csv"))
}
}
