% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-h5.R
\name{read_h5_counts}
\alias{read_h5_counts}
\title{Read scRNA-seq counts matrix in h5 format.
Automatically 'infer' counts;features;cells from the h5 file structure.
Attempts multiple ways.}
\usage{
read_h5_counts(h5.file)
}
\arguments{
\item{h5.file}{Path to an h5 / h5ad file.}
}
\value{
List with \code{counts} (features x cells) and \code{samples}
  (cell metadata, or NULL).
}
\description{
Read scRNA-seq counts matrix in h5 format.
Automatically 'infer' counts;features;cells from the h5 file structure.
Attempts multiple ways.
}
