% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spectronaut.R
\name{read_spectronaut}
\alias{read_spectronaut}
\title{Read Spectronaut output abundance global proteome data file.}
\usage{
read_spectronaut(file)
}
\arguments{
\item{file}{Path to Spectronaut output abundance file.}
}
\value{
abundance data matrix (features on rows; samples on columns)
}
\description{
Read Spectronaut output abundance global proteome data file.
}
