% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_counts}
\alias{validate_counts}
\title{Validate counts data}
\usage{
validate_counts(df)
}
\arguments{
\item{data}{matrix.}
}
\value{
boolean. true if data is valid
}
\description{
Counts data is valid if:
 - no duplicate rows
 - no empty rows
 - no duplicate cols
}
